requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore==1.7.3
google-cloud-storage==1.13.2
google-cloud-vision==0.35.2
